+++ src/Light.cpp
@@ -1,6 +1,6 @@
 
 #include "Light.h"
 
-Light::Light(glm::vec3 pos, glm::vec3 color)
+Light::Light(glm::dvec3 pos, glm::dvec3 color)
 	: pos(pos), color(color)
 {}
